forward: Prevent repeated memory allocation for the same tensor
When executing the VM op in a forward direction, the input tensor is
copied to produce a new output tensor using CopyTensorValueWithNewId().
If the `input_tensor->device_address` is empty during copy, the
`output_tensor->device_address` will also be empty. As a result,
the two tensors cannot share the same device memory, leading to the
issue of repeated memory allocation for the same tensor.

This can be fixed by pre-allocating device memory for tensors with
empty device addresses before copying the new tensor.