forward: Fix sparse tensor shape infer error
When executing the vm op, the forward executor avoids the issue of repeated
device memory allocation by pre-creating device addresses. However, the creation
of device addresses is not complete and does not adapt well to dynamic shape
scenarios.

Reproducer:
  $ pytest est_csr_add.py::test_function_csr_add

  C++ Call Stack:
  ---
  #0 CheckInteger()
  #1 SparseMatrixAddInferShape() at mindspore/ops/infer/sparse_matrix_add.cc:96
  [...]

  ValueError: For primitive[SparseMatrixAdd], the A indptr must be equal to 1,
  but got 0.

When do sparse tensors add, the initialization of the sparse tensor calls the
StopGradient op and triggers the creation of device addresses. This creation is
based on CreateDeviceAddress(), but the resulting kernel tensor does not have
its address_common field assigned. Since the executor considers the sparse tensor
to be of dynamic shape, it triggers a two-phase inference. In the second phase,
based on the kernel tensor, the address_common is empty, which leads to an empty
shape for the inferred a_indptr, finally causing the add operation to fail.

Fix this by replacing CreateDeviceAddress() to CreateOutputTensorAddress() to
create device_address since address_common field of kernel_tensor will be
assigned in it.